requirements.txt
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.51.0
httplib2==0.19.0
jwcrypto==0.7
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==8.1.1
pycryptodome==3.9.7
pyocr==0.7.2
pytesseract==0.3.3
python-jwt==3.2.6
requests-toolbelt==0.9.1
sseclient==0.0.26